+++ src/registers/mod.rs
@@ -3,6 +3,7 @@
 pub mod control;
 pub mod model_specific;
 pub mod rflags;
+pub mod segmentation;
 pub mod xcontrol;
 
 #[cfg(feature = "instructions")]
